Bug #489294 The local time will overflow after long time bootup
[root cause  ] tv_ptr->sec is a uint32 data, for sec is ok. But when 
switch to us, it will overflow for about just about one and a half hour.
[changes     ] Change the data type to uint64 before switch.
[side effects] No
[self test   ]<9830aef, passed on long time call>:
[reviewers   ] wenmei.hua; Chengli.Du
[change_type ] feature_bugfix
[tag_product ] common

+++ video/video_2_0/jbv/jbv.c
@@ -120,7 +120,7 @@ static void _JBV_putPkt(
         obj_ptr->firstAtime    = pkt_ptr->atime;
         obj_ptr->firstUnNormTs = pkt_ptr->tsOrig;
         updateFirstTs = 1;
-        obj_ptr->initTime  = tv_ptr->sec * _JBV_SEC_TO_USEC + tv_ptr->usec;
+        obj_ptr->initTime  = ((uint64)tv_ptr->sec) * _JBV_SEC_TO_USEC + (uint64)tv_ptr->usec;
         obj_ptr->tsLast90K = pkt_ptr->tsOrig;
         obj_ptr->ready     = 1;
 
@@ -221,7 +221,7 @@ static void _JBV_putPkt(
                               firstSeqnNextFrame,
                               lastSeqnNextFrame,
                               obj_ptr->unit[firstSeqnNextFrame].ts,
-                              tv_ptr->sec * _JBV_SEC_TO_USEC + tv_ptr->usec);
+                              ((uint64)tv_ptr->sec) * _JBV_SEC_TO_USEC + (uint64)tv_ptr->usec);
         }else{
             break;
         }
@@ -324,7 +324,7 @@ static void _JBV_getPkt(
         /*
          * Find time stamp of current time.
          */
-        curTime = tv_ptr->sec * _JBV_SEC_TO_USEC + tv_ptr->usec;
+        curTime = ((uint64)tv_ptr->sec) * _JBV_SEC_TO_USEC + (uint64)tv_ptr->usec;
         curTime -= obj_ptr->initTime;
 
         /* Draw rule bases on JBV state */
